Разбивка проекта на классы
