+++ README.md
@@ -44,6 +44,7 @@ key=15, cnt words = 47
 
 #### Как работает Статистический анализ (отношение количества
 ####глассных букв к согласным) работает далеко не всегда
+        0. Работает с криптографическим ключом = 9
         1. Вычисление отношения количества гласных букв к согласным в 
            дополнительном файле этого же автора
         2. Нахождение отношения количества гласных букв к согласным в
